fix: separate CHartAll component and fix unable to show pie chart
